Break mime document into a separate module
