+++ packages/docregistry/src/index.ts
@@ -3,6 +3,7 @@
 
 export * from './context';
 export * from './default';
+export * from './mimedocument';
 export * from './registry';
 
 import '../style/index.css';
